pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "interpretable-splade-classifier"
version = "2.0.0"
requires-python = ">=3.10"
dependencies = [
    "captum>=0.6.0",
    "datasets>=2.14.0",
    "lime>=0.2.0",
    "nltk>=3.8.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "torch>=2.1.0",
    "tqdm>=4.60.0",
    "transformers>=4.30.0",
]

[tool.pytest.ini_options]
markers = ["slow: tests requiring network access (deselect with '-k not slow')"]

[tool.setuptools.packages.find]
include = ["splade*"]
